+++ backend/src/main/java/com/timesnewronan/hoopiq_api/controller/PlayerController.java
@@ -2,6 +2,7 @@ package com.timesnewronan.hoopiq_api.controller;
 
 import com.timesnewronan.hoopiq_api.entity.Player;
 import com.timesnewronan.hoopiq_api.entity.PlayerSeasonStat;
+import com.timesnewronan.hoopiq_api.entity.PlayerGameStat;
 // import com.timesnewronan.hoopiq_api.repository.PlayerRepository;
 import com.timesnewronan.hoopiq_api.service.PlayerService;
 
@@ -57,5 +58,11 @@ public class PlayerController {
     @GetMapping("/{id}/season-stats")
     public List<PlayerSeasonStat> getSeasonStatsByPlayerId(@PathVariable Long id) {
         return playerService.getSeasonStatsByPlayerId(id);
+    }
+
+    // Request for player's games
+    @GetMapping("/{id}/games")
+    public List<PlayerGameStat> getPlayerGameStatsByPlayerId(@PathVariable Long id) {
+        return playerService.getPlayerGameStatsByPlayerId(id);
     }
 }