+++ sql/001_init_schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE players (
 );
 
 CREATE TABLE player_season_stats (
-    id BIGSERIAL PRIMARY KEY,
+    id ,SERIAL PRIMARY KEY,
     player_id BIGINT NOT NULL REFERENCES players(id),
     season VARCHAR(20) NOT NULL,
     games_played INTEGER,
@@ -39,4 +39,36 @@ CREATE TABLE player_season_stats (
     free_throw_pct NUMERIC(5,3),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE (player_id, season)
+);
+
+CREATE TABLE games (
+    id BIGSERIAL PRIMARY KEY,
+    nba_game_id INTEGER NOT NULL UNIQUE,
+    home_team_id BIGINT NOT NULL REFERENCES teams(id),
+    away_team_id BIGINT NOT NULL REFERENCES teams(id),
+    home_team_score INTEGER NOT NULL,
+    away_team_score INTEGER NOT NULL,
+    game_date DATE NOT NULL,
+    season VARCHAR(20) NOT NULL,
+    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+);
+
+CREATE TABLE player_game_logs (
+    id BIGSERIAL PRIMARY KEY,
+    player_id BIGINT NOT NULL REFERENCES players(id),
+    game_id BIGINT NOT NULL REFERENCES games(id),
+    team_id BIGINT NOT NULL REFERENCES teams(id),
+    pts_scored INTEGER NOT NULL,
+    assists INTEGER NOT NULL,
+    rebounds INTEGER NOT NULL, 
+    blocks INTEGER NOT NULL,
+    steals INTEGER NOT NULL,
+    minutes NUMERIC(5,2),
+    field_goal_pct NUMERIC(5,3),
+    three_point_pct NUMERIC(5,3),
+    free_throw_pct NUMERIC(5,3),
+    turnovers INTEGER NOT NULL,
+    win_loss CHAR(1),
+    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP, 
+    UNIQUE (player_id, game_id)
 );